requirements.txt
flask
spacy
pymupdf
google-generativeai
